Add logging for all update methods
This should make updates a lot safer, but it is incredibly ugly. I want
to do this with a decorator, but python is being annoying so I will
revisit. Now that we are getting new users I just want this
functionality in the bot before I want to spend the time to learn the
proper decorator pattern.